Rename monitoringStatusValueFormatters to monitoringItemsValueFormatters
