+++ utils/monitoringItemsValueFormatters.ts
